Реализована обработка заказов на покупку телеграм звезд.
Заказы для выполнения теперь объединяются в одну пачку. Благодаря этому, если возникла проблема с каким-то заказом, это не останавливает поток выполнения.

Обработка ошибок во время оформления заказов с отправкой уведомления админу

Проверка баланса для выполнения стандартных заказов перенесена в соотвествующий метод. Ранее проверка баланса была до того, как все заказы передавались в работу.

Отображение заказов:
Заказы на Телеграм звезды игнорируются в выдаче непринятых, активных и архивных заказов

Локализованы тексты сообщений хендлера покупки телеграм звезд

Исправлен баг с оплатой заказа после его удаления. Теперь это два взаимоисключающих действия

Переименована общая коллекция для хранения всех новых заказов
Поправлено отображение сумм в сообщениях. Теперь число всегда имеет 2 цифры после точки.
Убраны лишние принты

+++ src/handlers/buy_stars/st1_intro.py
@@ -3,7 +3,9 @@ from aiogram.filters import Command
 from aiogram.fsm.context import FSMContext
 from aiogram.fsm.storage.base import StorageKey
 
+import config
 from core.db import users
+from core.localisation.texts import messages
 from core.localisation.texts.buttons import buy_stars_button
 from core.storage import storage
 from loader import bot, dp
@@ -12,10 +14,13 @@ from utils.callback_templates import balance_recharge_template
 
 
 
-text = 'Введи количество звезд: От 50 до 1000000 (1 млн)'
+
 @dp.callback_query(F.data == buy_stars_button.get('callback'))
 async def handle_payment(query: types.CallbackQuery, state: FSMContext):
-
+    user_id = query.from_user.id
+    if user_id != config.ADMIN_ID:
+        await query.answer('Will available soon')
+        return
     await cmd_handler(query.message, state)
     await query.answer()
     await query.message.delete()
@@ -25,6 +30,9 @@ async def handle_payment(query: types.CallbackQuery, state: FSMContext):
 @dp.message(Command('buy_stars'))
 async def cmd_handler(msg: types.Message, state: FSMContext):
     user_id = msg.chat.id
+    if user_id != config.ADMIN_ID:
+        return
+
     key = StorageKey(bot.id, user_id, user_id)
     data = await storage.get_data(key)
     msgs_to_delete = data.get('msgs_to_delete', [])
@@ -32,6 +40,6 @@ async def cmd_handler(msg: types.Message, state: FSMContext):
     lang = users.get_user_lang(user_id)
 
     await state.set_state(None)
-    answer = await msg.answer(text)
+    answer = await msg.answer(messages.tg_stars_enter_quantity[lang])
     await storage.update_data(key, msgs_to_delete=msgs_to_delete + [answer.message_id])
     await state.set_state(states.BuyStars.waiting_for_amount)
